requirements.txt
# Core dependencies for SMS Spam Detection
streamlit==1.28.1
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3

# Text processing
nltk==3.8.1
regex==2023.8.8

# Deep learning (optional - for DistilBERT)
torch==2.0.1
transformers==4.33.2

# API dependencies
flask==2.3.3
flask-cors==4.0.0

# Data validation
pydantic==2.3.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Configuration
pyyaml==6.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Logging
loguru==0.7.0

# Environment management
python-dotenv==1.0.0

# Progress bars
tqdm==4.66.1
